added test.rs, implemented PR#317 review
